# src/main/java/com/example/fraud/schema/SchemaRequest.java
package com.example.fraud.schema;

import java.util.List;

public record SchemaRequest(
    String eventType,
    String displayName,
    String description,
    List<SchemaFieldDefinition> fields
) {}